Corrections for Init
