+++ src/main/java/test/tutorial/dragoon/springboot/learningspringboot/controller/TestController.java
@@ -2,7 +2,7 @@ package test.tutorial.dragoon.springboot.learningspringboot.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import test.tutorial.dragoon.springboot.learningspringboot.InitDto;
+import test.tutorial.dragoon.springboot.learningspringboot.dto.InitDto;
 
 
 @RestController
