--- src/test/java/test/tutorial/dragoon/springboot/learningspringboot/LearningSpringbootApplicationTests.java
@@ -1,17 +1,0 @@
-package test.tutorial.dragoon.springboot.learningspringboot;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class LearningSpringbootApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
-
